Fix #103, add option to configure base tick rate
SCH_LAB was fixed at 1Hz because it bound to the 1Hz tick from CFE_TIME.
This creates an OSAL timer instead, which posts a semaphore, and this
can serve as a much more flexible time source, with configurable rate.

This still binds to 1Hz and the SCH does not start its first message
until the first 1Hz is received, thereby keeping some form of 1Hz sync.